vacancyController

+++ src/main/java/com/example/jobsearch/model/Vacancy.java
@@ -1,19 +1,21 @@
 package com.example.jobsearch.model;
 
+import lombok.Builder;
 import lombok.Data;
 
 import java.time.LocalDateTime;
 @Data
+
 public class Vacancy {
-    private long id;
+    private Long id;
     private String title;
     private String description;
-    private int categoryId;
+    private Long categoryId;
     private double salary;
     private int expFrom;
     private int expTo;
     private boolean isActive;
-    private int employerId;
+    private Long employerId;
     private LocalDateTime createdDate;
     private LocalDateTime updateDate;
 }
